修改了cinema_pages里的adminx，models和views，增加了基于用户打分的个性化推荐（基于统计的，即从用户最近的打分最高的8个评论对应的电影，得到8*2=16部推荐（可能少于16部，这里设置少于8部就用默认的）对应StaRecom）

+++ apps/movie/views.py
@@ -65,7 +65,7 @@ class AddReview(View):
             movie_comments.movie = movie
             movie_comments.content = comments
             movie_comments.user = request.user
-            movie_comments.star = int(star)
+            movie_comments.star = float(star)
             # movie_comments.user_id = 1
             # movie_comments.movie_id = 2
             # movie_comments.content = 'hello'
